requirements.txt
aiohttp==3.11.11
fastapi==0.115.8
jinja2==3.1.5
openai==1.60.2
pydantic==2.10.6
python-dotenv==1.0.1
typing-extensions==4.12.2
uvicorn==0.34.0
